+++ database/migrations/2024_09_09_141600_add_membergroup_to_events.php
@@ -26,7 +26,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('events', function (Blueprint $table) {
-            $table->dropColumn('member_group_id');
+            $table->dropConstrainedForeignId('member_group_id');
 
             $table->boolean('logged_in_only')->default(false)->after('enabled');
         });